added error pages

+++ python/mysite/production/restart_servers.sh
@@ -2,12 +2,20 @@
 
 # Needs sudo permission.
 # This scripts restarts nginx and uwsgi servers.
+# setup_servers.sh does the initial setup of the servers.
 
 # STEP 1: restart uwsgi servers
 
 echo "Restarting uwsgi...";
-killall uwsgi;
-sleep 4;
+if [[ -e runtime/uwsgi.pid ]]; then
+  /usr/local/bin/uwsgi --stop runtime/uwsgi.pid;
+  #/usr/local/bin/uwsgi --reload runtime/uwsgi.pid;
+else
+  echo "runtime/uwsgi.pid file not found. Going for a brutal `killall uwsgi`!";
+  killall uwsgi;
+  killall uwsgi;
+fi
+sleep 2;
 /usr/local/bin/uwsgi --emperor /etc/uwsgi/vassals --uid www-data --gid www-data --daemonize /var/log/uwsgi-emperor.log;
 
 # STEP 2: restart nginx server
